bossbar fix on world load/change. New %skip-cause% placeholder for bo…

+++ src/main/java/me/mrgeneralq/sleepmost/services/SleepService.java
@@ -102,7 +102,7 @@ public class SleepService implements ISleepService {
 
     @Override
     public double getSleepersPercentage(World world) {
-        return getSleepersAmount(world) / getPlayerCountInWorld(world);
+        return (double)getSleepersAmount(world) / (double)getPlayerCountInWorld(world);
     }
 
     @Override
